requirements.txt
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
transformers
torch
pyenchant
